Move firebase init to its own file
