+++ src/front-end/src/components/MentorshipForm.js
@@ -4,22 +4,7 @@ import { Form, Button, Input, DatePicker, Tooltip, Select } from 'antd';
 import { InfoCircleOutlined, QuestionCircleOutlined } from '@ant-design/icons';
 import TimezonePicker from 'react-bootstrap-timezone-picker';
 import 'react-bootstrap-timezone-picker/dist/react-bootstrap-timezone-picker.min.css';
-
-import firebase from 'firebase/app';
-import 'firebase/auth';
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: 'AIzaSyAdutyXDo_0GvanZ58V1l702Co_zQu4a3M',
-  authDomain: 'mentormatch-afa46.firebaseapp.com',
-  databaseURL: 'https://mentormatch-afa46.firebaseio.com',
-  projectId: 'mentormatch-afa46',
-  storageBucket: 'mentormatch-afa46.appspot.com',
-  messagingSenderId: '515592331593',
-  appId: '1:515592331593:web:c38e0d3defadf9886cfa98',
-};
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
+import firebase from '../firebase-init';
 
 const { RangePicker } = DatePicker;
