+++ src/front-end/src/firebase-init.js
@@ -1,0 +1,17 @@
+import firebase from 'firebase/app';
+import 'firebase/auth';
+
+// Your web app's Firebase configuration
+const firebaseConfig = {
+  apiKey: 'AIzaSyAdutyXDo_0GvanZ58V1l702Co_zQu4a3M',
+  authDomain: 'mentormatch-afa46.firebaseapp.com',
+  databaseURL: 'https://mentormatch-afa46.firebaseio.com',
+  projectId: 'mentormatch-afa46',
+  storageBucket: 'mentormatch-afa46.appspot.com',
+  messagingSenderId: '515592331593',
+  appId: '1:515592331593:web:c38e0d3defadf9886cfa98',
+};
+// Initialize Firebase
+firebase.initializeApp(firebaseConfig);
+
+export default firebase;
